requirements.txt
boto3
pandas
pyspark>=3.2.0
pyarrow
snowflake-connector-python
sqlalchemy
pyyaml
apache-airflow==2.6.3
pytest
